requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
starlette==0.37.0
pymongo==4.6.0
motor==3.3.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic[email]==2.9.0
email-validator==2.2.0
